fix(frontend): update how react app loads, including main.tsx
