--- src/frontend/src/index.ts
@@ -1,1 +1,0 @@
-import('./App');
