--- src/frontend/src/sum.js
@@ -1,3 +1,0 @@
-export function sum(a, b) {
-  return a + b;
-}
