+++ src/frontend/tests/App.test.jsx
@@ -1,5 +1,8 @@
-import { sum } from '../src/sum';
 import { describe, expect, test } from 'vitest';
+
+function sum(a, b) {
+  return a + b;
+}
 
 describe('Test', () => {
   test('adds 1 + 2 to equal 3', () => {
